movie DB 저장 완료 / Schedule 자동 제작 코드 구현중

+++ MobleMovieProject/src/main/java/com/movie/vo/TheaterVo.java
@@ -8,7 +8,6 @@ public class TheaterVo {
 	private CinemaVo cinema_id;
 	private CinemaVo cinema_name;
 	private String theater_name;
-	private Integer theater_seat;
-	private Integer theater_price;
-
+	private Long theater_seat;
+	private Long theater_price;
 }
